feat(do-xang-online): park-first refuel flow and split station scene
Add parking animation, session composables, and station UI primitives.

Replace PumpScene with FuelStationScene and subcomponents.

Car off-screen until park; responsive motion; reduced-motion fix.
