+++ src/views/do-xang-online/index.vue
@@ -1,22 +1,19 @@
 <script setup lang="ts">
 import { computed, onMounted, ref, watch } from 'vue'
 import { RouterLink } from 'vue-router'
-import { useClipboard, useIntervalFn, useMediaQuery, useRafFn } from '@vueuse/core'
+import { useClipboard, useMediaQuery } from '@vueuse/core'
 import meta from './meta'
 import { useFuelRetailPrices } from './composables/useFuelRetailPrices'
+import { useParkingAnimation } from './composables/useParkingAnimation'
+import { useRefuelSession } from './composables/useRefuelSession'
+import { stationStatusLabel, useRefuelUiState } from './composables/useRefuelStationMachine'
 import PriceBoard from './components/PriceBoard.vue'
 import VehicleCascadePicker from './components/VehicleCascadePicker.vue'
 import FuelSelector from './components/FuelSelector.vue'
-import PumpScene from './components/PumpScene.vue'
+import FuelStationScene from './components/FuelStationScene.vue'
 import ReceiptPanel from './components/ReceiptPanel.vue'
 import { DEFAULT_VEHICLE_ID, EMPTY_VEHICLE_ID, resolveVehicle } from './data/vehicleCatalog'
-import type { FillPhaseCode } from './types'
-import {
-  SESSION_COUPLE_MS,
-  SESSION_HOSE_PRIME_MS,
-  SESSION_PREP_MS,
-  SESSION_WALK_MS,
-} from './sessionConstants'
+import { SESSION_WALK_MS } from './sessionConstants'
 
 type FuelDef = {
   id: string
@@ -61,37 +58,46 @@ const canPump = computed(
 )
 const selectedFuel = computed(() => fuels.value.find((f) => f.id === selectedFuelId.value)!)
 
-onMounted(() => {
-  fetchPrices()
-})
-
-const hoseSessionKey = ref(0)
-
-const fillingState = ref<'idle' | 'filling' | 'complete'>('idle')
-const liters = ref(0)
-const displayedLiters = ref(0)
-const pulseArmed = ref(false)
-const sessionStartedAt = ref<number | null>(null)
-const sessionElapsedMs = ref(0)
-const hoseLinePercent = ref(0)
-const attendantAtVehicle = ref(false)
-const fillPhaseCode = ref<FillPhaseCode>('idle')
+const capacityLiters = computed(() => selectedVehicle.value.capacityLiters)
 
 const prefersReducedMotion = useMediaQuery('(prefers-reduced-motion: reduce)')
 
-/** Thời điểm bắt đầu tính lít / % bình — sau khi ống đã đầy (prep + đi + cắm + chảy ống) */
-function sessionPreTankMs() {
-  if (prefersReducedMotion.value) return 0
-  return SESSION_PREP_MS + SESSION_WALK_MS + SESSION_COUPLE_MS + SESSION_HOSE_PRIME_MS
-}
+const { isParked, isParking, parkingProgress01, beginParking, resetParking } = useParkingAnimation({
+  prefersReducedMotion,
+})
 
-const minFillMs = 5000
-const maxFillMs = 10000
-const fillDurationMs = computed(() => {
-  const cap = selectedVehicle.value.capacityLiters
-  if (cap <= 0) return minFillMs
-  const ms = 4000 + cap * 120
-  return Math.max(minFillMs, Math.min(maxFillMs, ms))
+const {
+  hoseSessionKey,
+  fillingState,
+  liters,
+  displayedLiters,
+  pulseArmed,
+  sessionElapsedMs,
+  hoseLinePercent,
+  attendantAtVehicle,
+  fillPhaseCode,
+  startFill,
+  stopFill,
+  resetFill: resetSessionOnly,
+  resetSessionCore,
+} = useRefuelSession({
+  capacityLiters,
+  prefersReducedMotion,
+})
+
+const refuelUiState = useRefuelUiState({
+  isParked,
+  isParking,
+  fillingState,
+  fillPhaseCode,
+})
+
+const getStationStateLabel = computed(() =>
+  stationStatusLabel(refuelUiState.value, fillPhaseCode.value, fillingState.value),
+)
+
+onMounted(() => {
+  fetchPrices()
 })
 
 const progressPercent = computed(() => {
@@ -146,158 +152,9 @@ const receiptText = computed(() => {
 
 const { copy, copied } = useClipboard()
 
-const getStationStateLabel = computed(() => {
-  if (fillingState.value === 'idle') return 'Chờ bắt đầu'
-  if (fillingState.value === 'complete') return 'Đã đầy bình'
-  switch (fillPhaseCode.value) {
-    case 'prep':
-      return 'Lấy vòi tại trụ...'
-    case 'walk':
-      return 'Kéo ống tới xe...'
-    case 'couple':
-      return 'Đang cắm vòi vào xe...'
-    case 'hose':
-      return 'Xăng đang chảy trong ống...'
-    case 'tank':
-      return 'Xăng đang vào bình xe...'
-    default:
-      return 'Đang đổ...'
-  }
-})
-
-function applyFillingTick() {
-  if (fillingState.value !== 'filling') return
-  if (sessionStartedAt.value === null) return
-
-  const cap = selectedVehicle.value.capacityLiters
-  if (cap <= 0) return
-
-  const now = performance.now()
-  const t = now - sessionStartedAt.value
-  const pre = sessionPreTankMs()
-
-  if (prefersReducedMotion.value) {
-    attendantAtVehicle.value = true
-    hoseLinePercent.value = 100
-    fillPhaseCode.value = 'tank'
-  } else if (t < SESSION_PREP_MS) {
-    attendantAtVehicle.value = false
-    hoseLinePercent.value = 0
-    fillPhaseCode.value = 'prep'
-    liters.value = 0
-    return
-  } else if (t < SESSION_PREP_MS + SESSION_WALK_MS) {
-    attendantAtVehicle.value = false
-    hoseLinePercent.value = 0
-    fillPhaseCode.value = 'walk'
-    liters.value = 0
-    return
-  } else if (t < SESSION_PREP_MS + SESSION_WALK_MS + SESSION_COUPLE_MS) {
-    attendantAtVehicle.value = true
-    hoseLinePercent.value = 0
-    fillPhaseCode.value = 'couple'
-    liters.value = 0
-    return
-  } else {
-    attendantAtVehicle.value = true
-    const hoseStart = SESSION_PREP_MS + SESSION_WALK_MS + SESSION_COUPLE_MS
-    if (t < pre) {
-      fillPhaseCode.value = 'hose'
-      const u = Math.max(0, Math.min(1, (t - hoseStart) / SESSION_HOSE_PRIME_MS))
-      // smoothstep: đầu/cuối chảy chậm hơn — cảm giác xăng “lấp ống” từ từ
-      const eased = u * u * (3 - 2 * u)
-      hoseLinePercent.value = Math.min(100, eased * 100)
-      liters.value = 0
-      return
-    }
-    hoseLinePercent.value = 100
-    fillPhaseCode.value = 'tank'
-  }
-
-  const tankElapsed = t - pre
-  const ratio = Math.max(0, Math.min(1, tankElapsed / fillDurationMs.value))
-  liters.value = cap * ratio
-
-  if (ratio >= 1) {
-    pause()
-    fillingState.value = 'complete'
-    fillPhaseCode.value = 'done'
-  }
-}
-
-const { pause, resume } = useIntervalFn(applyFillingTick, 80, { immediate: false })
-
 watch([selectedVehicleId, selectedFuelId], () => {
-  pause()
-  liters.value = 0
-  displayedLiters.value = 0
-  fillingState.value = 'idle'
-  pulseArmed.value = false
-  sessionStartedAt.value = null
-  sessionElapsedMs.value = 0
-  hoseLinePercent.value = 0
-  attendantAtVehicle.value = false
-  fillPhaseCode.value = 'idle'
-})
-
-const { pause: pauseDisplay, resume: resumeDisplay } = useRafFn(
-  ({ delta }) => {
-    if (fillingState.value === 'filling' && sessionStartedAt.value !== null) {
-      sessionElapsedMs.value = performance.now() - sessionStartedAt.value
-    } else if (fillingState.value !== 'filling') {
-      sessionElapsedMs.value = 0
-    }
-
-    if (prefersReducedMotion.value) return
-    if (fillingState.value !== 'filling') return
-
-    if (!pulseArmed.value && sessionStartedAt.value !== null) {
-      const now = performance.now()
-      const t = now - sessionStartedAt.value
-      if (t >= sessionPreTankMs() + 400) pulseArmed.value = true
-    }
-
-    const cap = selectedVehicle.value.capacityLiters
-    const target = Math.min(liters.value, cap)
-    const diff = target - displayedLiters.value
-
-    if (Math.abs(diff) < 0.0005) {
-      displayedLiters.value = target
-      return
-    }
-
-    const alpha = 1 - Math.exp(-delta / 180)
-    displayedLiters.value += diff * alpha
-  },
-  { immediate: false },
-)
-
-watch(
-  fillingState,
-  (state) => {
-    if (prefersReducedMotion.value) {
-      pauseDisplay()
-      displayedLiters.value =
-        state === 'complete' ? selectedVehicle.value.capacityLiters : liters.value
-      return
-    }
-
-    if (state === 'filling') {
-      resumeDisplay()
-      return
-    }
-
-    pauseDisplay()
-    displayedLiters.value =
-      state === 'complete' ? selectedVehicle.value.capacityLiters : liters.value
-  },
-  { immediate: true },
-)
-
-watch(liters, (value) => {
-  if (!prefersReducedMotion.value) return
-  if (fillingState.value !== 'filling') return
-  displayedLiters.value = value
+  resetSessionCore()
+  resetParking()
 })
 
 const dropSeeds = computed(() => {
@@ -317,58 +174,22 @@ const dropSeeds = computed(() => {
   })
 })
 
-function startFill() {
+function handlePark() {
   if (!canPump.value) return
-  if (fillingState.value === 'complete') return
-  if (fillingState.value === 'idle') {
-    liters.value = 0
-    displayedLiters.value = 0
-  }
-
-  hoseSessionKey.value += 1
-
-  const now = performance.now()
-  sessionStartedAt.value = now
-  pulseArmed.value = false
-  if (prefersReducedMotion.value) {
-    attendantAtVehicle.value = true
-    hoseLinePercent.value = 100
-    fillPhaseCode.value = 'tank'
-    pulseArmed.value = true
-  } else {
-    attendantAtVehicle.value = false
-    hoseLinePercent.value = 0
-    fillPhaseCode.value = 'prep'
-  }
-  fillingState.value = 'filling'
-  resume()
-  applyFillingTick()
+  if (isParking.value || isParked.value) return
+  beginParking()
 }
 
-function stopFill() {
-  if (fillingState.value !== 'filling') return
-  pause()
-  fillingState.value = 'idle'
-  displayedLiters.value = liters.value
-  pulseArmed.value = false
-  sessionStartedAt.value = null
-  sessionElapsedMs.value = 0
-  hoseLinePercent.value = 0
-  attendantAtVehicle.value = false
-  fillPhaseCode.value = 'idle'
+function handleStartRefuel() {
+  if (!canPump.value) return
+  if (!isParked.value) return
+  if (fillingState.value === 'filling') return
+  startFill()
 }
 
 function resetFill() {
-  pause()
-  liters.value = 0
-  displayedLiters.value = 0
-  fillingState.value = 'idle'
-  pulseArmed.value = false
-  sessionStartedAt.value = null
-  sessionElapsedMs.value = 0
-  hoseLinePercent.value = 0
-  attendantAtVehicle.value = false
-  fillPhaseCode.value = 'idle'
+  resetSessionOnly()
+  resetParking()
 }
 
 async function handleCopyReceipt() {
@@ -441,7 +262,7 @@ async function handleCopyReceipt() {
           />
         </div>
         <div class="lg:col-span-7 animate-fade-up animate-delay-6">
-          <PumpScene
+          <FuelStationScene
             :vehicle="selectedVehicle"
             :can-pump="canPump"
             :fuel="{ label: selectedFuel.label, pricePerLiter: selectedFuel.pricePerLiter }"
@@ -462,7 +283,11 @@ async function handleCopyReceipt() {
             :is-using-fallback="isUsingFallback"
             :liters-odometer-key="litersOdometerKey"
             :cost-odometer-key="costOdometerKey"
-            @start-fill="startFill"
+            :is-parked="isParked"
+            :is-parking="isParking"
+            :parking-progress01="parkingProgress01"
+            @park="handlePark"
+            @start-refuel="handleStartRefuel"
             @stop-fill="stopFill"
             @reset-fill="resetFill"
           />
